+++ site/docs/pages/api/build-swap-transaction.mdx
@@ -115,8 +115,8 @@ const response = await buildSwapTransaction({
 
 ## Returns
 
-[`Promise<BuildSwapTransactionResponse>`](/swap/types#buildswaptransactionresponse)
+[`Promise<BuildSwapTransactionResponse>`](/api/types#buildswaptransactionresponse)
 
 ## Parameters
 
-[`BuildSwapTransactionParams`](/swap/types#buildswaptransactionparams)
+[`BuildSwapTransactionParams`](/api/types#buildswaptransactionparams)
