+++ site/docs/pages/api/get-swap-quote.mdx
@@ -70,8 +70,8 @@ const quote = await getSwapQuote({
 
 ## Returns
 
-[`Promise<GetSwapQuoteResponse>`](/swap/types#getswapquoteresponse)
+[`Promise<GetSwapQuoteResponse>`](/api/types#getswapquoteresponse)
 
 ## Parameters
 
-[`GetSwapQuoteParams`](/swap/types#getswapquoteparams)
+[`GetSwapQuoteParams`](/api/types#getswapquoteparams)
