+++ site/docs/pages/api/get-tokens.mdx
@@ -98,8 +98,8 @@ const tokens = await getTokens({
 
 ## Returns
 
-[`Promise<GetTokensResponse>`](/token/types#gettokensresponse)
+[`Promise<GetTokensResponse>`](/api/types#gettokensresponse)
 
 ## Parameters
 
-[`GetTokensOptions`](/token/types#gettokensoptions)
+[`GetTokensOptions`](/api/types#gettokensoptions)
