+++ src/api/buildSwapTransaction.test.ts
@@ -1,14 +1,14 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { CDP_GET_SWAP_TRADE } from '../../network/definitions/swap';
-import { sendRequest } from '../../network/request';
-import { DEGEN_TOKEN, ETH_TOKEN } from '../mocks';
-import type { BuildSwapTransaction } from '../types';
-import { buildSwapTransaction } from './buildSwapTransaction';
+import { CDP_GET_SWAP_TRADE } from '../network/definitions/swap';
+import { sendRequest } from '../network/request';
+import { DEGEN_TOKEN, ETH_TOKEN } from '../swap/mocks';
+import type { BuildSwapTransaction } from '../swap/types';
 /**
  * @vitest-environment node
  */
-import { getAPIParamsForToken } from './getAPIParamsForToken';
-import { getSwapTransaction } from './getSwapTransaction';
+import { getAPIParamsForToken } from '../swap/utils/getAPIParamsForToken';
+import { buildSwapTransaction } from './buildSwapTransaction';
+import { getSwapTransaction } from './utils/getSwapTransaction';
 
 vi.mock('../../network/request');
