+++ src/api/types.ts
@@ -1,4 +1,8 @@
+import type { Address } from 'viem';
+import type { SwapQuote } from '../swap/types';
 import type { Token } from '../token/types';
+
+export type AddressOrETH = Address | 'ETH';
 
 /**
  * Note: exported as public Type
@@ -8,6 +12,48 @@ export type APIError = {
   error: string; // The Error long message
   message: string; // The Error short message
 };
+
+/**
+ * Note: exported as public Type
+ */
+export type BuildSwapTransactionParams = GetSwapQuoteParams & {
+  fromAddress: Address; // The address of the user
+};
+
+export type GetAPIParamsForToken =
+  | GetSwapQuoteParams
+  | BuildSwapTransactionParams;
+
+export type GetQuoteAPIParams = {
+  amount: string; // The amount to be swapped
+  amountReference?: string; // The reference amount for the swap
+  from: AddressOrETH | ''; // The source address or 'ETH' for Ethereum
+  to: AddressOrETH | ''; // The destination address or 'ETH' for Ethereum
+  v2Enabled?: boolean; // Whether to use V2 of the API (default: false)
+  slippagePercentage?: string; // The slippage percentage for the swap
+};
+
+export type GetSwapAPIParams = GetQuoteAPIParams & {
+  fromAddress: Address; // The address of the user
+};
+
+/**
+ * Note: exported as public Type
+ */
+export type GetSwapQuoteParams = {
+  amount: string; // The amount to be swapped
+  amountReference?: string; // The reference amount for the swap
+  from: Token; // The source token for the swap
+  isAmountInDecimals?: boolean; // Whether the amount is in decimals
+  maxSlippage?: string; // The slippage of the swap
+  to: Token; // The destination token for the swap
+  useAggregator: boolean; // Whether to use a DEX aggregator
+};
+
+/**
+ * Note: exported as public Type
+ */
+export type GetSwapQuoteResponse = SwapQuote | APIError;
 
 /**
  * Note: exported as public Type
@@ -22,3 +68,5 @@ export type GetTokensOptions = {
  * Note: exported as public Type
  */
 export type GetTokensResponse = Token[] | APIError;
+
+export type SwapAPIParams = GetQuoteAPIParams | GetSwapAPIParams;
