+++ src/swap/types.ts
@@ -4,8 +4,6 @@ import type { Config } from 'wagmi';
 import type { SendTransactionMutateAsync } from 'wagmi/query';
 import type { Token } from '../token/types';
 
-export type AddressOrETH = Address | 'ETH';
-
 /**
  * Note: exported as public Type
  */
@@ -25,53 +23,11 @@ export type BuildSwapTransactionResponse = BuildSwapTransaction | SwapError;
 /**
  * Note: exported as public Type
  */
-export type BuildSwapTransactionParams = GetSwapQuoteParams & {
-  fromAddress: Address; // The address of the user
-};
-
-/**
- * Note: exported as public Type
- */
 export type Fee = {
   amount: string; // The amount of the fee
   baseAsset: Token; // The base asset for the fee
   percentage: string; // The percentage of the fee
 };
-
-export type GetAPIParamsForToken =
-  | GetSwapQuoteParams
-  | BuildSwapTransactionParams;
-
-export type GetQuoteAPIParams = {
-  amount: string; // The amount to be swapped
-  amountReference?: string; // The reference amount for the swap
-  from: AddressOrETH | ''; // The source address or 'ETH' for Ethereum
-  to: AddressOrETH | ''; // The destination address or 'ETH' for Ethereum
-  v2Enabled?: boolean; // Whether to use V2 of the API (default: false)
-  slippagePercentage?: string; // The slippage percentage for the swap
-};
-
-export type GetSwapAPIParams = GetQuoteAPIParams & {
-  fromAddress: Address; // The address of the user
-};
-
-/**
- * Note: exported as public Type
- */
-export type GetSwapQuoteParams = {
-  amount: string; // The amount to be swapped
-  amountReference?: string; // The reference amount for the swap
-  from: Token; // The source token for the swap
-  isAmountInDecimals?: boolean; // Whether the amount is in decimals
-  maxSlippage?: string; // The slippage of the swap
-  to: Token; // The destination token for the swap
-  useAggregator: boolean; // Whether to use a DEX aggregator
-};
-
-/**
- * Note: exported as public Type
- */
-export type GetSwapQuoteResponse = SwapQuote | SwapError;
 
 export type GetSwapMessageParams = {
   error?: SwapError;
@@ -155,8 +111,6 @@ export type SwapAmountInputReact = {
   token?: Token; // Selected token
   type: 'to' | 'from'; // Identifies if component is for toToken or fromToken
 };
-
-export type SwapAPIParams = GetQuoteAPIParams | GetSwapAPIParams;
 
 export type SwapAPIResponse = {
   approveTx?: RawTransactionData; // The approval transaction
